gpu_requirements.txt
numpy
scipy
matplotlib
seaborn
pandas
networkx
tensorflow-gpu
h5py
scikit-learn
gensim
pyyaml
keras